refactor: #12 댓글 CRUD API 수정
